Use NewEmptySnapshotStorageQuotaReporter in main, use separate main for tests
