+++ cloud/disk_manager/internal/pkg/facade/disk_service_test/ya.make
@@ -8,7 +8,12 @@ SET_APPEND(RECIPE_ARGS --with-cells)
 SET_APPEND(RECIPE_ARGS --encryption)
 SET_APPEND(RECIPE_ARGS --creation-and-deletion-allowed-only-for-disks-with-id-prefix "Test")
 SET_APPEND(RECIPE_ARGS --disable-disk-registry-based-disks)
+SET_APPEND(RECIPE_ARGS --disk-manager-binary-path cloud/disk_manager/test/mocks/disk-manager/disk-manager-test)
 INCLUDE(${ARCADIA_ROOT}/cloud/disk_manager/internal/pkg/facade/testcommon/common.inc)
+
+DEPENDS(
+    cloud/disk_manager/test/mocks/disk-manager
+)
 
 GO_XTEST_SRCS(
     disk_relocation_test.go
