+++ cloud/disk_manager/pkg/snapshot/ya.make
@@ -2,7 +2,6 @@ GO_LIBRARY()
 
 SRCS(
     quota.go
-    quota_stub.go
 )
 
 END()
