+++ cloud/disk_manager/test/mocks/disk-manager/ya.make
@@ -1,0 +1,7 @@
+GO_PROGRAM(disk-manager-test)
+
+SRCS(
+    main.go
+)
+
+END()
