+++ cloud/disk_manager/test/mocks/ya.make
@@ -1,5 +1,6 @@
 RECURSE(
     compute
+    disk-manager
     kms
     metadata
     s3-quota-proxy
